Fixed build not pass.
